# .factory/api-contracts.yaml
openapi: 3.0.3
info:
  title: Dom'Atic API
  description: |
    Local IoT Hub — centralized device management, real-time monitoring,
    automation rules, and security alerting. All endpoints (except auth)
    require a valid Bearer JWT token obtained via POST /auth/login.
  version: 1.0.0
  contact:
    name: Dom'Atic
    email: admin@domatic.local
servers:
  - url: http://domatic.local:8000
    description: Local LAN server
  - url: http://localhost:8000
    description: Development

# ---------------------------------------------------------------------------
# Security Schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived access JWT (15 min). Obtain via POST /auth/login.
        Refresh via POST /auth/refresh using the HttpOnly cookie.

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    PaginatedMeta:
      type: object
      properties:
        total:    { type: integer, example: 142 }
        page:     { type: integer, example: 1 }
        per_page: { type: integer, example: 50 }
        pages:    { type: integer, example: 3 }

    # --- Error ---
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  loc:  { type: array, items: { type: string } }
                  msg:  { type: string }
                  type: { type: string }

    # --- Enums (shared) ---
    ProtocolType:
      type: string
      enum: [wifi, zigbee, zwave, ble, matter, ethernet, unknown]
    ChipType:
      type: string
      enum: [esp8266, esp32, esp32s2, esp32s3, esp32c3, bk7231, bk7231n, rtl8710, unknown]
    FirmwareType:
      type: string
      enum: [tasmota, esphome, openbk, openshock, original, custom, unknown]
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, format: password }
    TokenResponse:
      type: object
      properties:
        access_token:  { type: string }
        token_type:    { type: string, example: bearer }
        expires_in:    { type: integer, description: Seconds until expiry, example: 900 }
    UserMe:
      type: object
      properties:
        id:           { type: string, format: uuid }
        username:     { type: string }
        email:        { type: string, format: email }
        is_superuser: { type: boolean }
        last_login:   { type: string, format: date-time, nullable: true }

    # --- Rooms ---
...
    Device:
      type: object
      properties:
        id:                        { type: string, format: uuid }
        name:                      { type: string }
        room_id:                   { type: string, format: uuid, nullable: true }
        room_name:                 { type: string, nullable: true }
        protocol:                  { $ref: '#/components/schemas/ProtocolType' }
        chip_type:                 { $ref: '#/components/schemas/ChipType' }
        ip_address:                { type: string, format: ipv4, nullable: true }
        mac_address:               { type: string, example: 'AA:BB:CC:DD:EE:FF', nullable: true }
        firmware_type:             { $ref: '#/components/schemas/FirmwareType' }
        firmware_version:          { type: string, nullable: true }
        firmware_latest:           { type: string, nullable: true }
        firmware_update_available: { type: boolean }
        mqtt_topic:                { type: string, nullable: true }
        mqtt_lwt_topic:            { type: string, nullable: true }
        is_online:                 { type: boolean }
        is_cloud_dependent:        { type: boolean }
        is_blocked:                { type: boolean }
        blocked_reason:            { type: string, nullable: true }
        cloud_attempts:            { type: integer }
        last_seen:                 { type: string, format: date-time, nullable: true }
        metadata:                  { type: object, additionalProperties: true }
        unresolved_security_events: { type: integer }
        created_at:                { type: string, format: date-time }
        updated_at:                { type: string, format: date-time }
    DeviceCreate:
      type: object
      required: [name]
      properties:
        name:              { type: string, maxLength: 150 }
        room_id:           { type: string, format: uuid, nullable: true }
        protocol:          { $ref: '#/components/schemas/ProtocolType' }
        chip_type:         { $ref: '#/components/schemas/ChipType' }
        ip_address:        { type: string, format: ipv4, nullable: true }
        mac_address:       { type: string, nullable: true }
        firmware_type:     { $ref: '#/components/schemas/FirmwareType' }
        firmware_version:  { type: string, nullable: true }
        mqtt_topic:        { type: string, nullable: true }
        mqtt_lwt_topic:    { type: string, nullable: true }
        is_cloud_dependent: { type: boolean, default: false }
        metadata:          { type: object, additionalProperties: true, default: {} }
    DeviceUpdate:
      type: object
      properties:
        name:              { type: string, maxLength: 150 }
        room_id:           { type: string, format: uuid, nullable: true }
        protocol:          { $ref: '#/components/schemas/ProtocolType' }
        chip_type:         { $ref: '#/components/schemas/ChipType' }
        ip_address:        { type: string, format: ipv4, nullable: true }
        mac_address:       { type: string, nullable: true }
        firmware_type:     { $ref: '#/components/schemas/FirmwareType' }
        firmware_version:  { type: string, nullable: true }
        firmware_latest:   { type: string, nullable: true }
        mqtt_topic:        { type: string, nullable: true }
        mqtt_lwt_topic:    { type: string, nullable: true }
        is_cloud_dependent: { type: boolean }
        is_blocked:        { type: boolean }
        blocked_reason:    { type: string, nullable: true }
        metadata:          { type: object, additionalProperties: true }
    DeviceState:
      type: object
      properties:
        device_id:   { type: string, format: uuid }
        state:       { type: object, additionalProperties: true }
        mqtt_topic:  { type: string, nullable: true }
        recorded_at: { type: string, format: date-time }
    DeviceCommand:
      type: object
      required: [command]
      properties:
        command:
          type: string
          enum: [power_on, power_off, power_toggle, restart, query_state, custom_mqtt]
          description: >
            power_on/off/toggle — sends MQTT command to cmnd/{topic}/POWER.
            restart — sends cmnd/{topic}/Restart 1.
            query_state — sends cmnd/{topic}/Status 0.
            custom_mqtt — sends payload to custom_topic.
        custom_topic:   { type: string, description: Required when command=custom_mqtt }
        custom_payload: { type: string, description: Required when command=custom_mqtt }

    # --- Security ---
...
    Automation:
      type: object
      properties:
        id:           { type: string, format: uuid }
        name:         { type: string }
        description:  { type: string, nullable: true }
        enabled:      { type: boolean }
        trigger:      { type: object, additionalProperties: true }
        conditions:   { type: array, items: { type: object } }
        actions:      { type: array, items: { type: object } }
        run_count:    { type: integer }
        last_triggered: { type: string, format: date-time, nullable: true }
        last_error:   { type: string, nullable: true }
        created_at:   { type: string, format: date-time }
        updated_at:   { type: string, format: date-time }
    AutomationCreate:
      type: object
      required: [name, trigger, actions]
      properties:
        name:        { type: string, maxLength: 150 }
        description: { type: string, nullable: true }
        enabled:     { type: boolean, default: true }
        trigger:
          type: object
          required: [type]
          additionalProperties: true
          example:
            type: mqtt_message
            topic: "stat/sonoff-01/POWER"
            value: "ON"
        conditions:
          type: array
          items: { type: object }
          default: []
        actions:
          type: array
          items: { type: object }
          minItems: 1
          example:
            - type: mqtt_publish
              topic: "cmnd/sonoff-02/POWER"
              payload: "ON"
    AutomationUpdate:
      type: object
      properties:
        name:        { type: string, maxLength: 150 }
        description: { type: string, nullable: true }
        enabled:     { type: boolean }
        trigger:     { type: object, additionalProperties: true }
        conditions:  { type: array, items: { type: object } }
        actions:     { type: array, items: { type: object } }
    AutomationTestResult:
      type: object
      properties:
        success:    { type: boolean }
        actions_run: { type: integer }
        results:
          type: array
          items:
            type: object
            properties:
              action: { type: object }
              success: { type: boolean }
              error:  { type: string, nullable: true }
              duration_ms: { type: number }

    # --- MQTT ---
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Obtain access + refresh tokens
      security: []                   # No auth required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refresh_token=eyJ...; HttpOnly; SameSite=Strict; Path=/auth/refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using HttpOnly cookie
      security: []
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Refresh token missing or expired
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMe'
        '401':
          description: Not authenticated
...
  /devices:
    get:
      tags: [Devices]
      summary: List all devices (with room join, no N+1)
      parameters:
        - in: query
          name: room_id
          schema: { type: string, format: uuid }
        - in: query
          name: protocol
          schema: { $ref: '#/components/schemas/ProtocolType' }
        - in: query
          name: firmware_type
          schema: { $ref: '#/components/schemas/FirmwareType' }
        - in: query
          name: is_online
          schema: { type: boolean }
        - in: query
          name: is_cloud_dependent
          schema: { type: boolean }
        - in: query
          name: search
          schema: { type: string, description: 'Trigram search on device name' }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Paginated device list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Device' }
                  meta:
                    $ref: '#/components/schemas/PaginatedMeta'

    post:
      tags: [Devices]
      summary: Register a new device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '409':
          description: MAC address or IP already registered
  /devices/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [Devices]
      summary: Get single device by ID
      responses:
        '200':
          description: Device detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Not found

    patch:
      tags: [Devices]
      summary: Update device fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        '200':
          description: Updated device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Not found

    delete:
      tags: [Devices]
      summary: Delete device and all related records
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /devices/{id}/state:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Devices]
      summary: Get most recent state snapshot for a device
      responses:
        '200':
          description: Latest device state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceState'
        '404':
          description: Device not found or no state recorded
  /devices/{id}/history:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Devices]
      summary: State history for a device (always requires time range for partition pruning)
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, format: date-time }
          description: Start of time range (ISO 8601). Required — enables partition pruning.
        - in: query
          name: to
          schema: { type: string, format: date-time }
          description: End of time range. Defaults to now.
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 1000 }
      responses:
        '200':
          description: State history list
          content:
            application/json:
              schema:
                type: object
                properties:
                  device_id:  { type: string, format: uuid }
                  from:       { type: string, format: date-time }
                  to:         { type: string, format: date-time }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DeviceState' }
  /devices/{id}/command:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Devices]
      summary: Send a command to the device via MQTT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCommand'
      responses:
        '200':
          description: Command dispatched
          content:
            application/json:
              schema:
                type: object
                properties:
                  queued:     { type: boolean }
                  mqtt_topic: { type: string }
                  payload:    { type: string }
        '400':
          description: Device has no MQTT topic configured
        '404':
          description: Device not found

  # =========================================================================
  # ROOMS
  # =========================================================================
...
  /automations:
    get:
      tags: [Automations]
      summary: List all automation rules
      parameters:
        - in: query
          name: enabled
          schema: { type: boolean }
      responses:
        '200':
          description: Automation list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Automation' }

    post:
      tags: [Automations]
      summary: Create a new automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationCreate'
      responses:
        '201':
          description: Automation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Automation'
  /automations/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [Automations]
      summary: Get automation detail
      responses:
        '200':
          description: Automation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Automation'

    patch:
      tags: [Automations]
      summary: Update automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationUpdate'
      responses:
        '200':
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Automation'

    delete:
      tags: [Automations]
      summary: Delete automation rule
      responses:
        '204':
          description: Deleted
  /automations/{id}/toggle:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Automations]
      summary: Toggle automation enabled/disabled
      responses:
        '200':
          description: Updated automation with new enabled state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Automation'
  /automations/{id}/test:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Automations]
      summary: Dry-run an automation (skips conditions, executes actions with dry_run=true flag)
      responses:
        '200':
          description: Test execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationTestResult'

  # =========================================================================
  # MQTT
  # =========================================================================
...